Decouple Function Curves From Grid Information
We don't really need the full ECLGraph structure in order to
extract the curves for scaled/unscaled saturation functions or
the FVF/viscosity property curves as functions of pressure. This
commit switches to using ECLInitFileData exclusively as the
source for those curves, albeit at the API cost of the caller
having to provide a grid ID--a string containing the grid name,
empty string for the global grid and LGR name for a local
grid--and a saturation region (SATNUM) for the saturation
functions.

In other words, this switches the "active cell" to being
relative to the particular grid instead of relative to the
ECLGraph.