Add Support for Extracting Imbibition Curve Data Points
This commit extends the existing saturation function curve data
point extraction function (ECLSaturationFunc::getSatFuncCurve())
to support points on the imbibition curve.  To this end, we add
the notion of a "curve set" which is either the primary drainage
curve or the imbibition curve.  The former uses regions defined
by the SATNUM array while the latter uses regions defined by the
IMBNUM array.  Moreover, we use the "I*" family of scaled end
points--e.g., ISGCR, IKRORW, ISWU--when computing points on the
scaled imbibition curves.

Finally, we remove the "satnum" parameter of getSatFuncCurve()
since this region index now depends on the curve set and it's
easier to ensure saturation region consistency from within the
implementation of getSatFuncCurve().